# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Sale and Purchase Platform with Online Auction System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  buyer: "buyer",
  seller: "seller",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  avatar?: string | null;
}

export type UpdateRoleBodyRole =
  (typeof UpdateRoleBodyRole)[keyof typeof UpdateRoleBodyRole];

export const UpdateRoleBodyRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export interface UpdateRoleBody {
  role: UpdateRoleBodyRole;
}

export type VehicleCondition =
  (typeof VehicleCondition)[keyof typeof VehicleCondition];

export const VehicleCondition = {
  new: "new",
  used: "used",
  certified_pre_owned: "certified_pre_owned",
} as const;

export type VehicleStatus = (typeof VehicleStatus)[keyof typeof VehicleStatus];

export const VehicleStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  sold: "sold",
} as const;

export interface Vehicle {
  id: number;
  sellerId: number;
  /** Present when sold via buy-now */
  buyerId?: number | null;
  sellerName: string;
  title: string;
  description: string;
  price: number;
  /** @nullable */
  buyNowPrice?: number | null;
  images: string[];
  make: string;
  model: string;
  year: number;
  /** @nullable */
  mileage?: number | null;
  condition: VehicleCondition;
  status: VehicleStatus;
  hasActiveAuction: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface VehicleListResponse {
  vehicles: Vehicle[];
  total: number;
}

export type CreateVehicleBodyCondition =
  (typeof CreateVehicleBodyCondition)[keyof typeof CreateVehicleBodyCondition];

export const CreateVehicleBodyCondition = {
  new: "new",
  used: "used",
  certified_pre_owned: "certified_pre_owned",
} as const;

export interface CreateVehicleBody {
  title: string;
  description: string;
  price: number;
  /** @nullable */
  buyNowPrice?: number | null;
  images: string[];
  make: string;
  model: string;
  year: number;
  /** @nullable */
  mileage?: number | null;
  condition: CreateVehicleBodyCondition;
}

export type UpdateVehicleBodyCondition =
  (typeof UpdateVehicleBodyCondition)[keyof typeof UpdateVehicleBodyCondition];

export const UpdateVehicleBodyCondition = {
  new: "new",
  used: "used",
  certified_pre_owned: "certified_pre_owned",
} as const;

export interface UpdateVehicleBody {
  title?: string;
  description?: string;
  price?: number;
  /** @nullable */
  buyNowPrice?: number | null;
  images?: string[];
  /** @nullable */
  mileage?: number | null;
  condition?: UpdateVehicleBodyCondition;
}

export type AuctionStatus = (typeof AuctionStatus)[keyof typeof AuctionStatus];

export const AuctionStatus = {
  upcoming: "upcoming",
  active: "active",
  ended: "ended",
  cancelled: "cancelled",
} as const;

export interface Auction {
  id: number;
  vehicleId: number;
  vehicle: Vehicle;
  startPrice: number;
  currentPrice: number;
  /** @nullable */
  reservePrice?: number | null;
  startTime: string;
  endTime: string;
  status: AuctionStatus;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
  bidCount: number;
  createdAt: string;
}

export type AuctionDetailStatus =
  (typeof AuctionDetailStatus)[keyof typeof AuctionDetailStatus];

export const AuctionDetailStatus = {
  upcoming: "upcoming",
  active: "active",
  ended: "ended",
  cancelled: "cancelled",
} as const;

export interface Bid {
  id: number;
  auctionId: number;
  userId: number;
  userName: string;
  amount: number;
  createdAt: string;
}

export interface AuctionDetail {
  id: number;
  vehicleId: number;
  vehicle: Vehicle;
  startPrice: number;
  currentPrice: number;
  /** @nullable */
  reservePrice?: number | null;
  startTime: string;
  endTime: string;
  status: AuctionDetailStatus;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
  bidCount: number;
  bids: Bid[];
  createdAt: string;
}

export interface AuctionListResponse {
  auctions: Auction[];
  total: number;
}

export interface CreateAuctionBody {
  vehicleId: number;
  startPrice: number;
  /** @nullable */
  reservePrice?: number | null;
  startTime: string;
  endTime: string;
}

export interface PlaceBidBody {
  amount: number;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  outbid: "outbid",
  auction_won: "auction_won",
  auction_ended: "auction_ended",
  vehicle_approved: "vehicle_approved",
  vehicle_rejected: "vehicle_rejected",
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  message: string;
  isRead: boolean;
  /** @nullable */
  auctionId?: number | null;
  /** @nullable */
  vehicleId?: number | null;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalVehicles: number;
  pendingVehicles: number;
  activeAuctions: number;
  totalBids: number;
  totalRevenue: number;
}

export type DashboardSummaryStats = {
  totalVehicles: number;
  activeAuctions: number;
  totalBids: number;
};

export interface DashboardSummary {
  featuredVehicles: Vehicle[];
  activeAuctions: Auction[];
  endingSoonAuctions: Auction[];
  stats: DashboardSummaryStats;
}

export type GetVehiclesParams = {
  search?: string;
  status?: GetVehiclesStatus;
  /**
   * @nullable
   */
  sellerId?: number | null;
  limit?: number;
  offset?: number;
};

export type GetVehiclesStatus =
  (typeof GetVehiclesStatus)[keyof typeof GetVehiclesStatus];

export const GetVehiclesStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type GetAuctionsParams = {
  status?: GetAuctionsStatus;
  limit?: number;
  offset?: number;
};

export type GetAuctionsStatus =
  (typeof GetAuctionsStatus)[keyof typeof GetAuctionsStatus];

export const GetAuctionsStatus = {
  active: "active",
  ended: "ended",
  upcoming: "upcoming",
} as const;