# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vehicle Sale and Purchase Platform with Online Auction System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auctions:
    get:
      operationId: getAuctions
      tags: [auctions]
      summary: List auctions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, ended, upcoming]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of auctions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionListResponse"
    post:
      operationId: createAuction
      tags: [auctions]
      summary: Create a new auction for a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuctionBody"
      responses:
        "201":
          description: Auction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "400":
          description: Vehicle not available for auction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auctions/{auctionId}:
    get:
      operationId: getAuction
      tags: [auctions]
      summary: Get auction by ID with bid history
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Auction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionDetail"
        "404":
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auctions/{auctionId}/bids:
    get:
      operationId: getAuctionBids
      tags: [bids]
      summary: Get all bids for an auction
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
    post:
      operationId: placeBid
      tags: [bids]
      summary: Place a bid on an auction
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBidBody"
      responses:
        "201":
          description: Bid placed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "400":
          description: Bid too low or auction not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users/{userId}/role:
    patch:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Admin - update user role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/vehicles/{vehicleId}/approve:
    patch:
      operationId: adminApproveVehicle
      tags: [admin]
      summary: Admin - approve vehicle listing
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /admin/vehicles/{vehicleId}/reject:
    patch:
      operationId: adminRejectVehicle
      tags: [admin]
      summary: Admin - reject vehicle listing
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [buyer, seller, admin]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    UpdateRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [buyer, seller, admin]
      required:
        - role
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        buyNowPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: ["integer", "null"]
        condition:
          type: string
          enum: [new, used, certified_pre_owned]
        status:
          type: string
          enum: [pending, approved, rejected, sold]
        hasActiveAuction:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - title
        - description
        - price
        - images
        - make
        - model
        - year
        - condition
        - status
        - hasActiveAuction
        - createdAt
...
    Auction:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        startPrice:
          type: number
        currentPrice:
          type: number
        reservePrice:
          type: ["number", "null"]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, active, ended, cancelled]
        winnerId:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
        bidCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - vehicleId
        - vehicle
        - startPrice
        - currentPrice
        - startTime
        - endTime
        - status
        - bidCount
        - createdAt
    AuctionDetail:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        startPrice:
          type: number
        currentPrice:
          type: number
        reservePrice:
          type: ["number", "null"]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, active, ended, cancelled]
        winnerId:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
        bidCount:
          type: integer
        bids:
          type: array
          items:
            $ref: "#/components/schemas/Bid"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - vehicleId
        - vehicle
        - startPrice
        - currentPrice
        - startTime
        - endTime
        - status
        - bidCount
        - bids
        - createdAt
    AuctionListResponse:
      type: object
      properties:
        auctions:
          type: array
          items:
            $ref: "#/components/schemas/Auction"
        total:
          type: integer
      required:
        - auctions
        - total
    CreateAuctionBody:
      type: object
      properties:
        vehicleId:
          type: integer
        startPrice:
          type: number
        reservePrice:
          type: ["number", "null"]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
      required:
        - vehicleId
        - startPrice
        - startTime
        - endTime
    Bid:
      type: object
      properties:
        id:
          type: integer
        auctionId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - auctionId
        - userId
        - userName
        - amount
        - createdAt
    PlaceBidBody:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount
...