# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Sale and Purchase Platform with Online Auction System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt: Date;
}